Dynamically select implementation
This will ensure always the right config will be picked up for the project, no matter if it's a bazel project or not.

Also, fixes a bug in BazelConfig where directory is never traversed up.
